rules: split up dep rules
This commit is a refactor which breaks up the source file containing all
the index and column dependency rules into an index rule file, a column
rule file, and an index-and-column rule file. This commit just moves
things around and fixes typos.

Release note: None
